requirements.txt
pymisp==2.4.134
PyYAML~=5.3.1
androguard==3.3.5
yara-python==3.11.0
